arm: DT: msm8226: Update min-default adreno frequency to be 200MHz
It's useless to stay at higher frequency by default, adreno is
efficient enough to render the 2D UI just fine at that frequency.

Higher default frequency means waste of power and, for environments
where power-efficiency is critical (applications such as handheld
battery-powered devices like smartphones), this is something that
should have be done a long time ago.

In any case, the governors are already deciding to use 200MHz, but
only when the GPU isn't _totally_ idle... and that's really poor,
because the lowest frequency wasn't being used for most of the uptime.

Update the "default" frequency (governor idling) to be 200MHz to get
the most battery power-saving, without the UI to feel sluggish at all.

Change-Id: I55b648de88d5c33b876215abef6578e0a7a75085
Signed-off-by: tarun93 <tarunmyid@gmail.com>